+++ docs/src/conf.py
@@ -101,6 +101,14 @@ nitpick_ignore = [
     ("py:class", "TaskStatus"),
     ("py:class", "ResultCode"),
     ("py:class", "ska_tango_base.type_hints.TaskFunctionType"),
+    # Ignored because of tangodiffdoc
+    ("py:class", "DevVarLongStringArray"),
+    ("py:class", "DevString"),
+    ("py:class", "DevUShort"),
+    ("py:class", "DevVarStringArray"),
+    ("py:class", "DevVoid"),
+    ("py:class", "DevULong"),
+
 ]
 
 # If your documentation needs a minimal Sphinx version, state it here.