+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+* [THORN-636] Added tests for unresponsive h/w.
 * [THORN-430] Added tangodocgen to the doc pipeline
 * [THORN-609] Improved health reporting docs.