+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 * [THORN-636] Added tests for unresponsive h/w.
+* [THORN-430] Added tangodocgen to the doc pipeline
 * [THORN-609] Improved health reporting docs.
 
 ## 7.1.0